gofumpt, rename package
