+++ hw09_struct_validator/validator/in_rule_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"github.com/vitermakov/otusgo-hw/hw09_struct_validator/validator"
+	"github.com/vitermakov/otusgo-hw/hw09structvalidator/validator"
 )
 
 func TestInRule(t *testing.T) {
